deleted 'plotting_unitary_events_analysis2.py &test_plotUE.py' because the weren't needed anymore;; renamed 'plotting_unitary_events_analysis.py to unitary_event_analysis.py' because all changes mentioned in #7 were made there. The former version of 'unitary_event_analysis.p' was afterwards deleted.
