+++ viziphant/unitary_event_analysis.py
@@ -2,6 +2,8 @@
 Plotting function for unitary event analysis results.
 """
 import math
+import string
+from collections import namedtuple
 
 import matplotlib.pyplot as plt
 import numpy as np
@@ -9,6 +11,10 @@ import quantities as pq
 from matplotlib.ticker import (MaxNLocator)
 
 import elephant.unitary_event_analysis as ue
+
+FigureUE = namedtuple(
+    "AxesUE", "spike_events, spike_rates, coincidence_events, "
+              "coincidence_rates, statistical_significance, unitary_events")
 
 params_dict_default = {
     # epochs to be marked on the time axis
@@ -24,31 +30,29 @@ params_dict_default = {
     # orientation (figure margins)
     'top': 0.95,
     'bottom': 0.1,
-    'right': 0.84,
+    'right': 0.92,
     'left': 0.08,
     # font size
     'fsize': 12,
     # the actual unit ids from the experimental recording
     'unit_real_ids': ['not specified', 'not specified'],
     # line width
-    'lw': 0.5,
+    'lw': 2,
     # y limit for the surprise
     'S_ylim': (-3, 3),
     # size of the marker
     'marker_size': 5,
     # size of the y-ticks interval for rasterplots
     'y_tick_interval': 15,
-    # show figure
-    'showfig': True,
-    # save figure
-    'savefig': False,
-    # path, file name and format for saving the figure
-    'path_filname_format': 'figure.pdf'
+    # uniform time unit
+    'time_unit': 'ms',
+    # uniform frequency unit
+    'frequency_unit': 'Hz',
 }
 
 
-def plot_UE(spiketrains, joint_surprise_dict, significance_level, bin_size,
-            window_size, window_step, **plot_params_user):
+def plot_unitary_events(data, joint_surprise_dict, significance_level, binsize,
+                        window_size, window_step, **plot_params_user):
     """
     Plots the results of unitary event analysis as a column of six subplots,
     comprised of raster plot, peri-stimulus time histogram, coincident event
@@ -57,12 +61,12 @@ def plot_UE(spiketrains, joint_surprise_dict, significance_level, bin_size,
 
     Parameters
     ----------
-    spiketrains : list of list of neo.SpikeTrain
-        A nested list of trails, neurons and their `neo.SpikeTrain` objects,
+    data : list of list of neo.SpikeTrain
+        A nested list of trails, neurons and there neo.SpikeTrain objects,
         respectively. This should be identical to the one used to generate
         joint_surprise_dict
     joint_surprise_dict : dict
-        The output of `elephant.unitary_event_analysis.jointJ_window_analysis`
+        The output of elephant.unitary_event_analysis.jointJ_window_analysis
         function. The values of each key has the shape of
             different pattern hash --> 0-axis
             different window --> 1-axis
@@ -81,24 +85,24 @@ def plot_UE(spiketrains, joint_surprise_dict, significance_level, bin_size,
     significance_level : float
         The significance threshold used to determine which coincident events
         are classified ad unitary events within a window.
-    bin_size : pq.Quantity
+    binsize : quantities.Quantity
         The size of bins for discretizing spike trains. This value should be
-        identical to the one used to generate `joint_surprise_dict`.
-    window_size : pq.Quantity
+        identical to the one used to generate joint_surprise_dict.
+    window_size : quantities.Quantity
         The size of the analysis-window. This value should be identical to the
-        one used to generate `joint_surprise_dict`.
-    window_step : pq.Quantity
+        one used to generate joint_surprise_dict.
+    window_step : quantities.Quantity
         The size of the window step. This value should be identical to th one
-        used to generate `joint_surprise_dict`.
-    **plot_params_user
-        Plotting parameters used to update the default plotting
+        used to generate joint_surprise_dict.
+    plot_params_user : dict
+        A dictionary of plotting parameters used to update the default plotting
         parameter values.
         Keys:
         -----
         events : dictionary (default: {})
             Epochs to be marked on the time axis
             key: epochs name as string
-            value: list of pq.Quantity
+            value: list of quantities.Quantity
         figsize : tuple of int (default: (12, 10))
             The dimensions for the figure size.
         hspace : float (default: 1)
@@ -122,34 +126,49 @@ def plot_UE(spiketrains, joint_surprise_dict, significance_level, bin_size,
             The marker size for the coincidence and unitary events.
         y_tick_interval : integers (default: 15)
             The size of the interval between y-ticks (for rasterplots only)
-        showfig : boolean (default: True)
-            Displays the figure on screen if True.
-        savefig : boolean (default: False)
-            Saves the figure to disk if True.
-        path_filname_format : string (default: figure.pdf)
-            The path and the filename to save the figure. The format is
-            inferred from the filename extension.
+        time_unit : string (default: 'ms')
+            The time unit used to rescale the spiketrains.
+        frequency_unit : string (default: 'Hz')
+            The frequency unit used to rescale the spikerates.
+    Returns
+    -------
+    result : instance of namedtuple()
+        The container for Axis objects generated by this function. Individual
+        axes can be accessed using the respective identifiers:
+        result.identifier
+        Identifiers: spike_events_axis, spike_rates_axis,
+                     coincidence_events_axis, coincidence_rates_axis,
+                     statistical_significance_axis, unitary_events_axis
     """
-
-    # # set common variables
-    n_neurons = len(spiketrains[0])
-    t_start = spiketrains[0][0].t_start
-    t_stop = spiketrains[0][0].t_stop
-    t_winpos = ue._winpos(t_start, t_stop, window_size, window_step)
-    n_trial = len(spiketrains)
-    joint_surprise_significance = ue.jointJ(significance_level)
-    xlim_left = (min(t_winpos)).rescale('ms').magnitude
-    xlim_right = (max(t_winpos) + window_size).rescale('ms').magnitude
-
     # update params_dict_default with user input
     params_dict = params_dict_default.copy()
     params_dict.update(plot_params_user)
 
+    # rescale all spiketrains to the uniform time unit from params_dict
+    for m in range(len(data)):
+        for n in range(len(data[0])):
+            data[m][n].rescale(params_dict['time_unit'])
+
+    # # set common variables
+    n_neurons = len(data[0])
+    t_start = data[0][0].t_start
+    t_stop = data[0][0].t_stop
+    t_winpos = ue._winpos(t_start, t_stop, window_size, window_step)
+    center_of_analysis_window = t_winpos + window_size / 2.
+    n_trials = len(data)
+    joint_surprise_significance = ue.jointJ(significance_level)
+    # print("min: ", min(t_winpos),
+    #       "max: ", max(t_winpos))
+    xlim_left = (min(t_winpos)).magnitude
+    xlim_right = (max(t_winpos) + window_size).magnitude
+    # print("xlim_left: ", xlim_left,
+    #       "xlim_right: ", xlim_right)
+
     if len(params_dict['unit_real_ids']) != n_neurons:
         raise ValueError(
-            f"The length of unit_ids should be equal to number of neurons! \n"
-            f"Unit_Ids ({params_dict['unit_real_ids']}) not equal number of "
-            f"neurons ({n_neurons})")
+            'length of unit_ids should be equal to number of neurons! \n'
+            'Unit_Ids: ' + params_dict['unit_real_ids']
+            + 'not equal number of neurons: ' + n_neurons)
 
     plt.figure(num=1, figsize=params_dict['figsize'])
     plt.subplots_adjust(hspace=params_dict['hspace'],
@@ -163,28 +182,29 @@ def plot_UE(spiketrains, joint_surprise_dict, significance_level, bin_size,
     y_tick_interval = params_dict['y_tick_interval']
     y_ticks_list = []
     # find start y-tick position for each trail
-    y_ticks_list.extend(range(1, n_neurons * n_trial, n_trial + 1))
+    for yt1 in range(1, n_neurons * n_trials, n_trials + 1):
+        y_ticks_list.append(yt1)
     # find in-between y-tick position for each trail with the specific interval
-    for neuron_id in range(n_neurons):
-        in_between_ticks = range(neuron_id * (n_trial + 1) + y_tick_interval,
-                                 (neuron_id + 1) * n_trial, y_tick_interval)
-        y_ticks_list.extend(in_between_ticks)
+    for n in range(n_neurons):
+        for yt2 in range(n * (n_trials + 1) + y_tick_interval,
+                         (n + 1) * n_trials, y_tick_interval):
+            y_ticks_list.append(yt2)
     y_ticks_list.sort()
     y_ticks_labels_list = [1]
     number_of_in_between_y_ticks_per_neuron = math.floor(
-        n_trial / y_tick_interval)
+        n_trials / y_tick_interval)
     for i in range(number_of_in_between_y_ticks_per_neuron):
         y_ticks_labels_list.append((i + 1) * y_tick_interval)
     auxiliary_list = y_ticks_labels_list
     # adding n_neuron times the y_ticks_labels_list to itself, so that each
     # neuron has the same y_ticks_labels
     for i in range(n_neurons - 1):
-        y_ticks_labels_list.extend(auxiliary_list)
+        y_ticks_labels_list += auxiliary_list
 
     def mark_epochs(axes_name):
         """
         Marks epochs on the respective axis by creating a vertical line and
-        showing the the epochs name under the last subplot. Epochs need to be
+        shows the epoch's name under the last subplot. Epochs need to be
         defined in the plot_params_user dictionary.
         Parameters
         ----------
@@ -196,103 +216,96 @@ def plot_UE(spiketrains, joint_surprise_dict, significance_level, bin_size,
                 axes_name.axvline(e_val, ls='-', lw=params_dict['lw'],
                                   color='r')
                 if axes_name.get_geometry()[2] == 6:
-                    axes_name.text(x=e_val - 10 * pq.ms, y=-65, s=key,
-                                   fontsize=9,
-                                   color='r')
+                    axes_name.text(x=e_val-10*pq.ms, y=-90, s=key, fontsize=12,
+                                   color='b')
 
     print('plotting Unitary Event Analysis ...')
 
     print('plotting Spike Events ...')
     axis1 = plt.subplot(6, 1, 1)
     axis1.set_title('Spike Events')
-    for neuron_id in range(n_neurons):
-        for trial, data_trial in enumerate(spiketrains):
-            spike_events_on_timescale = data_trial[neuron_id].rescale(
-                'ms').magnitude
+    for n in range(n_neurons):
+        for trial, data_trial in enumerate(data):
+            spike_events_on_timescale = data_trial[n].magnitude
             spike_events_on_trialscale = \
-                np.ones_like(data_trial[neuron_id].magnitude) * trial + \
-                neuron_id * (n_trial + 1) + 1
+                np.ones_like(data_trial[n].magnitude) * trial + \
+                n * (n_trials + 1) + 1
             axis1.plot(spike_events_on_timescale, spike_events_on_trialscale,
                        ls='none', marker='.', color='k', markersize=0.5)
-        if neuron_id < n_neurons - 1:
-            axis1.axhline((trial + 2) * (neuron_id + 1), lw=params_dict['lw'],
-                          color='b')
+        if n < n_neurons - 1:
+            axis1.axhline((trial + 2) * (n + 1), lw=params_dict['lw'],
+                          color='k')
     axis1.set_xlim(xlim_left, xlim_right)
-    axis1.set_ylim(0, (trial + 2) * (neuron_id + 1) + 1)
+    axis1.set_ylim(0, (trial + 2) * (n + 1) + 1)
     axis1.xaxis.set_major_locator(MaxNLocator(integer=True))
     axis1.set_yticks(y_ticks_list)
     axis1.set_yticklabels(y_ticks_labels_list, fontsize=params_dict['fsize'])
-    for neuron_id in range(n_neurons):
-        axis1.text(xlim_right + 20, neuron_id * (n_trial + 1),
-                   f"Unit {params_dict['unit_real_ids'][neuron_id]}")
+    axis1.text(xlim_right - 200, -50,
+               f"Unit {params_dict['unit_real_ids'][0]}")
+    axis1.text(xlim_right - 200, n_trials * n_neurons + 7,
+               f"Unit {params_dict['unit_real_ids'][1]}")
     axis1.set_ylabel('Trial', fontsize=params_dict['fsize'])
 
     print('plotting Spike Rates ...')
     axis2 = plt.subplot(6, 1, 2, sharex=axis1)
     axis2.set_title('Spike Rates')
-    # psth = peristimulu time histogram
+    # psth = peristimulus time histogram
     max_val_psth = 0
-    for neuron_id in range(n_neurons):
-        center_of_analysis_window = t_winpos + window_size / 2.
-        respective_rate_average = joint_surprise_dict['rate_avg'][:, neuron_id]
-        respective_rate_average = respective_rate_average.rescale('Hz')
+    for n in range(n_neurons):
+        respective_rate_average = joint_surprise_dict['rate_avg'][:, n].\
+            rescale(params_dict['frequency_unit'])
         axis2.plot(center_of_analysis_window, respective_rate_average,
-                   label=f"Unit {params_dict['unit_real_ids'][neuron_id]}",
+                   label=f"Unit {params_dict['unit_real_ids'][n]}",
                    lw=params_dict['lw'])
-        max_val_psth = max(*joint_surprise_dict['rate_avg'][:, neuron_id],
-                           max_val_psth)
+        if max(joint_surprise_dict['rate_avg'][:, n]) > max_val_psth:
+            max_val_psth = max(joint_surprise_dict['rate_avg'][:, n])
     axis2.set_xlim(xlim_left, xlim_right)
-    max_val_psth = max_val_psth.rescale('Hz').magnitude
-    axis2.set_ylim(0, max_val_psth + max_val_psth / 10)
+    max_val_psth = max_val_psth.rescale(params_dict['frequency_unit']).magnitude
+    axis2.set_ylim(0, max_val_psth + max_val_psth/10)
     axis2.xaxis.set_major_locator(MaxNLocator(integer=True))
     axis2.set_yticks([0, int(max_val_psth / 2), int(max_val_psth)])
-    mark_epochs(axis2)
-    axis2.legend(bbox_to_anchor=(1, 1), fancybox=True, shadow=True)
-    axis2.set_ylabel('(1/s)', fontsize=params_dict['fsize'])
+    axis2.legend(fontsize=params_dict['fsize']//2)
+    axis2.set_ylabel(f"({params_dict['frequency_unit']})",
+                     fontsize=params_dict['fsize'])
 
     print('plotting Coincidence Events ...')
     axis3 = plt.subplot(6, 1, 3, sharex=axis1)
     axis3.set_title('Coincidence Events')
-    for neuron_id in range(n_neurons):
-        for trial, data_trial in enumerate(spiketrains):
-            spike_events_on_timescale = data_trial[neuron_id].rescale(
-                'ms').magnitude
+    for n in range(n_neurons):
+        for trial, data_trial in enumerate(data):
+            spike_events_on_timescale = data_trial[n].magnitude
             spike_events_on_trialscale = \
-                np.ones_like(data_trial[neuron_id].magnitude) * trial + \
-                neuron_id * (n_trial + 1) + 1
+                np.ones_like(data_trial[n].magnitude) * trial + \
+                n * (n_trials + 1) + 1
             axis3.plot(spike_events_on_timescale, spike_events_on_trialscale,
                        ls='none', marker='.', color='k', markersize=0.5)
             coincidence_events_on_timescale = \
                 np.unique(joint_surprise_dict['indices']
-                          ['trial' + str(trial)]) * bin_size
+                          ['trial' + str(trial)]) * binsize
             coincidence_events_on_trialscale = \
                 np.ones_like(np.unique(joint_surprise_dict['indices'][
-                    'trial' + str(
-                        trial)])) * trial + neuron_id * (
-                    n_trial + 1) + 1
+                    'trial' + str(trial)])) * trial + n * (n_trials + 1) + 1
             axis3.plot(coincidence_events_on_timescale,
                        coincidence_events_on_trialscale, ls='',
                        markersize=params_dict['marker_size'], marker='s',
                        markerfacecolor='none', markeredgecolor='c')
-        if neuron_id < n_neurons - 1:
-            axis3.axhline((trial + 2) * (neuron_id + 1), lw=params_dict['lw'],
-                          color='b')
+        if n < n_neurons - 1:
+            axis3.axhline((trial + 2) * (n + 1), lw=params_dict['lw'],
+                          color='k')
     axis3.set_xlim(xlim_left, xlim_right)
-    axis3.set_ylim(0, (trial + 2) * (neuron_id + 1) + 1)
+    axis3.set_ylim(0, (trial + 2) * (n + 1) + 1)
     axis3.xaxis.set_major_locator(MaxNLocator(integer=True))
     axis3.set_yticks(y_ticks_list)
     axis3.set_yticklabels(y_ticks_labels_list, fontsize=params_dict['fsize'])
-    mark_epochs(axis3)
     axis3.set_ylabel('Trial', fontsize=params_dict['fsize'])
 
     print('plotting Coincidence Rates ..')
     axis4 = plt.subplot(6, 1, 4, sharex=axis1)
     axis4.set_title('Coincidence Rates')
-    center_of_analysis_window = t_winpos + window_size / 2.
     empirical_coincidence_rate = joint_surprise_dict['n_emp'] / \
-        (window_size.rescale('s').magnitude * n_trial)
+                                 (window_size.rescale('s').magnitude * n_trials)
     expected_coincidence_rate = joint_surprise_dict['n_exp'] / \
-        (window_size.rescale('s').magnitude * n_trial)
+                                (window_size.rescale('s').magnitude * n_trials)
     axis4.plot(center_of_analysis_window, empirical_coincidence_rate,
                label='empirical', lw=params_dict['lw'], color='c')
     axis4.plot(center_of_analysis_window, expected_coincidence_rate,
@@ -301,43 +314,37 @@ def plot_UE(spiketrains, joint_surprise_dict, significance_level, bin_size,
     axis4.xaxis.set_major_locator(MaxNLocator(integer=True))
     y_ticks = axis4.get_ylim()
     axis4.set_yticks([0, y_ticks[1] / 2, y_ticks[1]])
-    mark_epochs(axis4)
-    axis4.legend(bbox_to_anchor=(1, 1), fancybox=True, shadow=True)
-    axis4.set_ylabel('(1/s)', fontsize=params_dict['fsize'])
+    axis4.legend(fontsize=params_dict['fsize']//2)
+    axis4.set_ylabel(f"({params_dict['frequency_unit']})",
+                     fontsize=params_dict['fsize'])
 
     print('plotting Statistical Significance ...')
     axis5 = plt.subplot(6, 1, 5, sharex=axis1)
     axis5.set_title('Statistical Significance')
-    center_of_analysis_window = t_winpos + window_size / 2.
     joint_surprise_values = joint_surprise_dict['Js']
     axis5.plot(center_of_analysis_window, joint_surprise_values,
                lw=params_dict['lw'], color='k')
     axis5.set_xlim(xlim_left, xlim_right)
     axis5.set_ylim(params_dict['S_ylim'])
     axis5.axhline(joint_surprise_significance, ls='-', color='r')
-    axis5.axhline(-joint_surprise_significance, ls='-', color='b')
+    axis5.axhline(-joint_surprise_significance, ls='-', color='g')
     axis5.text(t_winpos[30], joint_surprise_significance + 0.3, '$\\alpha +$',
                color='r')
     axis5.text(t_winpos[30], -joint_surprise_significance - 0.9, '$\\alpha -$',
-               color='b')
+               color='g')
     axis5.xaxis.set_major_locator(MaxNLocator(integer=True))
-    axis5.set_yticks([ue.jointJ(0.99), ue.jointJ(0.5), ue.jointJ(0.01)])
-    mark_epochs(axis5)
-    axis5.set_yticklabels([0.99, 0.5, 0.01])
+    axis5.set_yticks([ue.jointJ(1-significance_level), ue.jointJ(0.5), ue.jointJ(0.01)])
+    axis5.set_yticklabels([1-significance_level, 0.5, 0.01])
 
-    print('plottting Unitary Events ...')
+    print('plotting Unitary Events ...')
     axis6 = plt.subplot(6, 1, 6, sharex=axis1)
     axis6.set_title('Unitary Events')
-    # sig_inx_win: indices of significant JointSurprises within a window
-    # x: indices of coincidence events in analysis window of a trail
-    # xx: indices_of_unitary_events_in_analysis_window_of_a_trial
-    for neuron_id in range(n_neurons):
-        for trial, data_trial in enumerate(spiketrains):
-            spike_events_on_timescale = data_trial[neuron_id].rescale(
-                'ms').magnitude
+    for n in range(n_neurons):
+        for trial, data_trial in enumerate(data):
+            spike_events_on_timescale = data_trial[n].magnitude
             spike_events_on_trialscale = \
-                np.ones_like(data_trial[neuron_id].magnitude) * trial + \
-                neuron_id * (n_trial + 1) + 1
+                np.ones_like(data_trial[n].magnitude) * trial + \
+                n * (n_trials + 1) + 1
             axis6.plot(spike_events_on_timescale, spike_events_on_trialscale,
                        ls='None', marker='.', markersize=0.5, color='k')
             indices_of_significant_JointSurprises = np.where(
@@ -349,10 +356,10 @@ def plot_UE(spiketrains, joint_surprise_dict, significance_level, bin_size,
                     indices_of_unitary_events = []
                     for j in indices_of_significant_JointSurprises:
                         coincidence_indices_greater_left_window_margin = \
-                            indices_of_coincidence_events * bin_size >= \
+                            indices_of_coincidence_events * binsize >= \
                             t_winpos[j]
                         coincidence_indices_smaller_right_window_margin = \
-                            indices_of_coincidence_events * bin_size < \
+                            indices_of_coincidence_events * binsize < \
                             t_winpos[j] + window_size
                         coincidence_indices_in_actual_analysis_window = \
                             coincidence_indices_greater_left_window_margin & \
@@ -362,30 +369,35 @@ def plot_UE(spiketrains, joint_surprise_dict, significance_level, bin_size,
                                 indices_of_unitary_events,
                                 indices_of_coincidence_events
                                 [coincidence_indices_in_actual_analysis_window]
-                            )
+                                )
                     unitary_events_on_timescale = \
-                        np.unique(indices_of_unitary_events) * bin_size
+                        np.unique(indices_of_unitary_events) * binsize
                     unitary_events_on_trialscale = \
                         np.ones_like(np.unique(indices_of_unitary_events)) * \
-                        trial + neuron_id * (n_trial + 1) + 1
+                        trial + n * (n_trials + 1) + 1
                     axis6.plot(unitary_events_on_timescale,
                                unitary_events_on_trialscale,
                                markersize=params_dict['marker_size'],
                                marker='s', ls='', markerfacecolor='none',
                                markeredgecolor='r')
-        if neuron_id < n_neurons - 1:
-            axis6.axhline((trial + 2) * (neuron_id + 1), lw=params_dict['lw'],
-                          color='b')
+        if n < n_neurons - 1:
+            axis6.axhline((trial + 2) * (n + 1), lw=params_dict['lw'],
+                          color='k')
     axis6.set_xlim(xlim_left, xlim_right)
-    axis6.set_ylim(0, (trial + 2) * (neuron_id + 1) + 1)
+    axis6.set_ylim(0, (trial + 2) * (n + 1) + 1)
     axis6.xaxis.set_major_locator(MaxNLocator(integer=True))
     axis6.set_yticks(y_ticks_list)
     axis6.set_yticklabels(y_ticks_labels_list, fontsize=params_dict['fsize'])
-    mark_epochs(axis6)
-    axis6.set_xlabel('Time [ms]', fontsize=params_dict['fsize'])
+    axis6.set_xlabel('Time (ms)', fontsize=params_dict['fsize'])
     axis6.set_ylabel('Trial', fontsize=params_dict['fsize'])
 
-    if params_dict['savefig']:
-        plt.savefig(params_dict['path_filename_format'])
-    if params_dict['showfig']:
-        plt.show()
+    # mark all epochs on all subplots and annotate all subplotaxis
+    for n in range(1, 7):
+        mark_epochs(eval(f"axis{n}"))
+        axis = locals()['axis'+str(n)]
+        axis.text(-0.05, 1.1, string.ascii_uppercase[n-1],
+                  transform=axis.transAxes, size=params_dict['fsize'] + 5,
+                  weight='bold')
+
+    result = FigureUE(axis1, axis2, axis3, axis4, axis5, axis6)
+    return result
